requirements.txt
beautifulsoup4==4.12.3
deepgram_sdk==3.6.0
google_api_python_client==2.146.0
groq==0.11.0
instaloader==4.13.1
Pillow==10.4.0
Requests==2.32.3
streamlit==1.37.1
